Add bashrc file via templating for Debian
Creates a Debian, Debian-slim, and Alpine template build
directory whose contents are copied to the docker build directories when
the corresponding Dockerfiles are generated by x.py.  In this way, the
files in each build-template directory can be further processed to
expand and replace macros.  And, the generated Dockerfiles can reference
these files, especially for ADD/COPY directives.